(refactor) Provide RecipeDatabase to decouple parsing and searching
The input (*.config and *.object files of the mod) has different
formats for extractors, smelters and centrifuges.

However, to keep our code clean,
we should have the same standartized "search in what we found" methods
for all recipes (regardless of whether it's extraction, crafting, etc.)

For that we provide add() method, which can be used by parsers
of *.config and *.object files to "register" found recipes in the DB.

Sanity checks (and Recipe class that does them) have been added to
ensure that incorrect recipes are not added into the database.