(cleanup) Remove unneeded code that was replaced by Cargo database
This wikitext is now gathered with {{All recipes for item}} template.
All we need to record is the Cargo database itself.

Note: not removing the RecipeSearchIndex class (for now) just in case
some degree of search would be needed in the future.
Its only method that is still being used is listKnownItems().