Seriously improve performance by caching the parsed JS: 12 -> 6 seconds
99,9% of the files are not modified between two runs of generate.js,
and it takes ~12 seconds to execute the JavaScript compiler to parse
all of them on every run.

We can reduce these 12 seconds to 6 seconds by saving an already parsed
script in a packed format (which is much faster to unpack than to run
the JavaScript compiler). Luckily it's pure JSON, not functional code.

We compare modification time of original script and of the .cache file,
and if the cache is newer than original file, then we use the cache.