Performance: cache results of "partitionKey -> chunkName" calculations
Determining the chunk via Consistent Hashing algorithm requires several
iterations and can be relatively slow (e.g. 6 seconds for 10 iterations
for everything the bot currently writes).

Reducing the number of iterations is not an option, as it increases the
chance of some chunks being much smaller than intended average,
and other chunks being much larger.

However, due to the very nature of Consistent Hashing, results of
these calculations are the same for the same key, unless there was a
change in the number of chunks. That means we can cache the results.