Remove icon= from "item" Cargo table: unused, increases size of chunks
We have a convention of uploading icons for item with id="SOMETHING"
as "File:Item icon SOMETHING.png". This makes a special field for that
superfluous, and unnecessarily increases the size of Cargo chunks.

It's not even useful, as this filename is relative to the directory in
which the .object/.item file resides, but we don't currently record
this directory for items.
Furthermore, MediaWiki wouldn't be able to extract icons from sprites
like "something.png:3", while our script (although it can't do now)
is theoretically capable of splitting those sprites in future versions.