ChunkWriter: add configuration variable to set "fixed number of chunks"
New variable: cargoChunksCount (129 is the current number of chunks).

This improves efficiency of Consistent Hashing
(we no longer need to reshuffle the keys unless the number of chunks
was explicitly modified in the configuration file).

The only reshuffling of keys happens in subchunks that were split due
to the "cargoMaxChunkSizeKBytes" variable.