Don't count header/footer as part of length of chunk, improve logging
Header and footer are lines like {{-start-}} for pywikibot,
they are not actually saved into the MediaWiki pages.
Therefore their length shouldn't count when splitting the chunk pages.

Additionally the debug.log will now show the lengths of all subchunks
(in situations when a chunk had to be split further due to the limit
of "cargoMaxChunkSizeKBytes" variable).