(cleanup) ResultsWriter: remove code that was replaced by use of Cargo
In the past [generate.js] generated wikitext of the page, and this
raw wikitext was written into the page. Removed methods were for that.

Now we just include some Cargo templates into the pages themselves,
while all the necessary information is in Cargo database
(which these templates can utilize to form the wikitext themselves).