(config) Exclude wizardspotionstation item (confusing crafting station)
This is incorrectly called "Alchemy Table" and crafts items from
Apothecary, causing recipes for Apothecary to be shown
as being crafted at the "Alchemy Table" (which is incorrect, because
there is a completely different crafting station "Alchemy Table").

We are excluding wizardspotionstation from scan, because this item
is unobtainable (is only in NPC dungeons).