(refactor) Remove RecipeSearchIndex - mostly unused
Methods like getRecipesWhereInputIs() were only needed when we didn't
use the Cargo database. They are no longer necessary, as we no longer
need to generate such lists for items (templates and Scribunto modules
can generate them on MediaWiki side with a simple SQL query to Cargo).

The only method that was still used - listKnownItems() - was moved into
the RecipeDatabase class. Also we no longer cache the results of this
method (what was previously RecipeSearchIndex.seenItems array),
because this method is only called once anyway.

This saves ~0.1 seconds per run of generate.js.