(refactor) Move icon-uploading code into [prepare_uploads.js] script
This saves ~0.3 seconds per run of generate.js.

We rarely need to update inventory icons, etc. (only when new items are
added into the game), and we need to run generate.js way more often.

Additionally, icons will now be searched in both the mod and vanilla,
because there are situations when Item from the mod refers to the icon
from vanilla.