(refactor) Provide AssetDatabase.shouldAdd(filename) - for linter, etc.
Some scripts other than generate.js might want to parse different lists
of assets. For example, only prepare_uploads.js needs .frames files,
and (hypothetical) linter (that we might write in the future)
would probably want to load ALL assets.

Because the assets are numerous and large, it's not a good idea to load
all of them (even those we don't need), and having shouldAdd() function
might allow us to configure "which assets to load" on per-script basis.

It comes with a performance cost of 0.3 seconds per non-cached load(),
probably due to internal optimizations of "fast-glob",
but most runs are cached and are not affected by this.