(refactor) Store treasure pools as RecipeSide, auto-compress RecipeSide
Elements in TreasurePoolDatabase.knownPools array are an item=>quantity
map, which is exactly the same thing as what RecipeSide represents.

Additionally, RecipeSide.compress() has been removed in favor of doing
the "can be merge these quantityAttributes into existing entry" check
in RecipeSide.addItem().

This doesn't result in any performance decrease, because those checks
are only performed if the same item is being added to RecipeSide twice.