(performance) Avoid overhead of require() in RecipeSide.toWikitext()
Despite the results of require() being cached, calling is still causes
each run of generate.js to be ~0.3 seconds slower compared to
using Query.findItem(), which remembers the ItemDatabase object.

Note: the reason why require() was even needed is because RecipeSide.js
is loaded before ItemDatabase (but toWikitext() is called after that).