Sort the results of listMentionedItemCodes() for consistency
Reason: the order of these results determines the order of Cargo rows
within the same chunk, and before this change it depended on order of
assets, which in turn depends on whether AssetDatabase was loaded from
cache or not.

Sometimes we need to manually remove the cache of AssetDatabase for
testing purposes (or to add a new asset type), and when that happened,
not having a consisent order of listMentionedItemCodes() caused
some Cargo rows to be reordered, leading to unnecessary edits.

Sorting the results of listMentionedItemCodes() mitigates the problem.