Split the Cargo chunks by the number of entries, not just by size in Kb
Main reason why saving pages with {{#cargo_store:}} directive causes
server load is the fact that DELETE/INSERT queries must be executed
for each such directive.

Since our goal is to reduce the load on the server, chunks should be
split by the number of {{#cargo_store:}} directives, not by the size
of Chunk page in kilobytes.

This is important for node/codex/armorset chunks, where 50Kb. page
could have contained a very small number of {{#cargo_store:}} rows.

We impose a reasonable limit (250 {{#cargo_store:}} per page, which
is close to what the heaviest Chunk pages had before this change,
and didn't cause any issues on the server with 20 seconds cooldown
between two bot edits).

The limit to length of Chunk page can now be relaxed to 200Kb.,
because MediaWiki is known to be able to handle this, and size has
significantly less impact on performance than number of rows.