(refactor) Pass optional renderParameters to RecipeComponent.toWikitext
Previously only the name of crafting station was passed (it was used to
add additional flags to {{CentrifugeRarity}} MediaWiki template).

Now we can pass other parameters (if any will be needed in the future),
and we no longer compare the crafting station name (string) on every
call of RecipeComponent.toWikitext(), which is called very often.