PageNameRegistry: call attemptRename on both (a, b) and (b, a)
This rules out the situation when some possibility to rename the item
has been skipped due to the order of candidates being suboptimal.

Since we sort the candidates by length of their itemCode/monster.type,
it's most likely (a, b) that will produce the renaming opportunity,
but we should check (b, a) as well just in case.