Support blueprint items ("<itemCode>-recipe") in RecipeComponent class
Many treasure pools (especially vanilla ones) drop blueprints,
and previously these pools were skipped, because there was no item
in the ItemDatabase with the IDs that these blueprints have.

Now we recognize the blueprints (e.g. "mmgravgun3-recipe")
and record them as "mmgravgun3" (unlocked item)
with { isBlueprint: true } in its quantityAttributes.

It's much cleaner than to create pseudo-items (and possibly articles)
for each of the blueprints.