Improve {{All recipes for pool}}: +TOC, +category, reorder
While having "Contents" on top of TreasurePool:<name> page makes sense,
having "How to obtain" on top makes it much easier to track the chain
between the desired item and the monsters that can drop it.

(especially since items have "How to obtain" on top of their pages)

So let's place "How to obtain" higher than "Contents" for convenience.