PlanetInfo: remove addColorToGravity (useless)
Very few planet types have the gravity override,
and even then the only ones with gravity >= 100 are Superdense/Shadow
(which we currently not even collecting for the Cargo database)
and Gas Giants (for which high gravity is obvious).

Also green color can be misinterpreted as "more convenient",
while in fact low gravity such as 20 (like on Moon) is not that good.