Remove "one of:" from spawntype lists: all listed monsters can spawn
While it's true that when the new monster spawns, it will be one of
the monsters listed in spawntype, it can be misinterpreted by reader
when inside the Recipe wikitext,
because "biome -> monster" recipes also show the "chance that this
monster will ever be chosen to appear on 1 planet".

If spawntype (say, Gleap+Fire Gleap) actually got chosen for a planet,
then both of those monsters can appear.

To avoid confusion, we are removing "One of:" header about such groups.