Provide biome infobox (split from recipes to allow human-written text)
We need to have screenshot and "on which planets" above the article,
or else the lengthy human-written description will result in those two
pieces of information to be somewhere at the bottom.

We have this problem with codex pages too, but codex pages are unlikely
to have human-written description. While biome pages sure can use it!

So the solution is to split the autogenerated into into infobox (on top
of the page) and recipes (at the bottom), same as on item pages.