Don't use GUI window titles for things other than multi-stage stations
Turns out, the GUI window titles are suboptimal for many items.

For example, some of them are in uppercase, some are less known than
the official name of the item, and some are too non-specific
(e.g. Apex Kitchen Counter has "Kitchen Counter" in its window title,
but Kitchen Counter is another item, so we need Apex in the item name).

Therefore we won't set shortdescription from GUI config EXCEPT FOR:
stages of multi-stage (upgradeable) crafting stations,
which don't really have a "shortdescription" of their own.