Dependencies: replace "strip-json-comments" with "node-json-minify"
Developer of "strip-json-comments" decided to break B/C for no reason,
so this package should be replaced as non-functional.

Not wasting my time on his "I love this new tech so much that you must
update to it now, or all my packages will break for you" blackmail.
There is 0 practical benefit today, and many highly qualified developers
(who honestly tried to update) have commented that this is a disaster.